backend-fast-poll-system/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
boto3==1.35.67
botocore==1.35.67
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.5
h11==0.14.0
idna==3.10
jmespath==1.0.1
load-dotenv==0.1.0
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.4
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1
